% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baseball}
\alias{baseball}
\title{Baseball hits}
\source{
\url{http://www.utstat.toronto.edu/~brunner/workshops/mixed/BinaryBballWithR.pdf}
}
\value{
\item{baseball}{a tibble}
}
\description{
Baseball hits
}
\details{
From website of Jerry Brunner:

"Right-handed basketball players take right and left-handed shots from 3
locations in a different random order for each player. Hit or miss is
recorded. This is a 2x3 factorial design with repeated measures on both
factors: Hand they are shooting with and spot on the court."
}
\examples{
data(baseball)
str(baseball)
}
\keyword{datasets}
