% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee.R
\name{gee_fit}
\alias{gee_fit}
\title{GEE fitting function}
\usage{
gee_fit(formula, data, family = gaussian, ...)
}
\arguments{
\item{formula}{Normal formula but uses the \code{gee_formula()} internal function for the \code{id_var()} specification for clustering}

\item{data}{Modeling data}

\item{family}{a family object: a list of functions and expressions for
defining link and variance functions. Families supported in gee are
gaussian, binomial, poisson, Gamma, and quasi; see the glm and family
documentation. Some links are not currently available: 1/mu^2 and sqrt have
not been hard-coded in the cgee engine at present. The inverse gaussian
variance function is not available. All combinations of remaining functions
can be obtained either by family selection or by the use of quasi.}

\item{...}{For additional parameters}
}
\description{
Custom fitting function to add GEE model with cluster
variable to parsnip GEE function call.
}
