% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hdp}
\alias{hdp}
\title{Simulated lung cancer data}
\source{
\url{https://stats.idre.ucla.edu/r/dae/mixed-effects-logistic-regression/},
\url{https://stats.idre.ucla.edu/stat/data/hdp.csv}
}
\value{
\item{hdp}{a tibble}
}
\description{
Simulated lung cancer data
}
\details{
A simulated data set with a binary outcome with levels "0" and "1".
Created by the UCLA Statistical Consulting group.
}
\examples{
data(hdp)
str(hdp)
}
\keyword{datasets}
