% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{longitudinal_counts}
\alias{longitudinal_counts}
\title{Simulated longitudinal Poisson counts}
\value{
\item{longitudinal_counts}{a tibble}
}
\description{
Simulated longitudinal Poisson counts
}
\details{
These are simulated data of 100 subjects each with 10 time points
and an additional numeric covariate. The linear predictor has a random
standard normal intercept per subject, a time coefficient of 1.50, and a
covariate coefficient of 0.25.
}
\examples{
data(longitudinal_counts)
str(longitudinal_counts)
}
\keyword{datasets}
