% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msa_data}
\alias{msa_data}
\title{Measurement systems analysis data}
\value{
\item{msa_data}{a tibble}
}
\description{
Measurement systems analysis data
}
\details{
A biological assay (i.e. a lab test) was run on 56 separate samples
twice. The goal is to measure what percentage of the total variation in the
results is related to the measurement system and how much is attributable to
the true systematic difference (sample-to-sample).
}
\examples{
data(msa_data)
str(msa_data)
}
\keyword{datasets}
