% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{riesby}
\alias{riesby}
\title{Imipramine longitudinal data}
\source{
Reisby, N., Gram, L.F., Bech, P. et al. Imipramine: Clinical effects
and pharmacokinetic variability. Psychopharmacology 54, 263-272 (1977).
}
\value{
\item{riesby}{a tibble}
}
\description{
Imipramine longitudinal data
}
\details{
These data are from a longitudinal clinical trial for depression.
The outcome is the change in depression scores week-to-week. The \code{endogenous}
column is an indicator for whether the subject fit the WHO Depression Scale
classification of endogenous. The \code{imipramine} and \code{desipramine} columns are
measurements of plasma levels for both substances.
}
\examples{
data(riesby)
str(riesby)
}
\keyword{datasets}
